third_party/WebKit/LayoutTests/fast/dom/global-event-handlers-expected.txt


FAIL No Element.onabort assert_false: expected false got true
PASS Set HTMLElement.onabort 
PASS Enumerate HTMLElement.onabort 
PASS Reflect HTMLElement.onabort 
PASS Set SVGElement.onabort 
PASS Enumerate SVGElement.onabort 
FAIL Reflect SVGElement.onabort assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onabort 
PASS Enumerate Document.onabort 
PASS Set Window.onabort 
PASS Enumerate Window.onabort 
FAIL No Element.onblur assert_false: expected false got true
PASS Set HTMLElement.onblur 
PASS Enumerate HTMLElement.onblur 
PASS Reflect HTMLElement.onblur 
PASS Set SVGElement.onblur 
PASS Enumerate SVGElement.onblur 
FAIL Reflect SVGElement.onblur assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onblur 
PASS Enumerate Document.onblur 
PASS Set Window.onblur 
PASS Enumerate Window.onblur 
FAIL No Element.onerror assert_false: expected false got true
PASS Set HTMLElement.onerror 
PASS Enumerate HTMLElement.onerror 
PASS Reflect HTMLElement.onerror 
PASS Set SVGElement.onerror 
PASS Enumerate SVGElement.onerror 
FAIL Reflect SVGElement.onerror assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onerror 
PASS Enumerate Document.onerror 
PASS Set Window.onerror 
PASS Enumerate Window.onerror 
FAIL No Element.onfocus assert_false: expected false got true
PASS Set HTMLElement.onfocus 
PASS Enumerate HTMLElement.onfocus 
PASS Reflect HTMLElement.onfocus 
PASS Set SVGElement.onfocus 
PASS Enumerate SVGElement.onfocus 
FAIL Reflect SVGElement.onfocus assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onfocus 
PASS Enumerate Document.onfocus 
PASS Set Window.onfocus 
PASS Enumerate Window.onfocus 
PASS No Element.oncancel 
FAIL Set HTMLElement.oncancel assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.oncancel assert_true: expected true got undefined
FAIL Reflect HTMLElement.oncancel assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.oncancel assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.oncancel assert_true: expected true got undefined
FAIL Reflect SVGElement.oncancel assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.oncancel assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.oncancel assert_true: expected true got undefined
FAIL Set Window.oncancel assert_equals: Initially null expected null but got undefined
FAIL Enumerate Window.oncancel assert_true: expected true got undefined
PASS No Element.oncanplay 
FAIL Set HTMLElement.oncanplay assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.oncanplay assert_true: expected true got undefined
FAIL Reflect HTMLElement.oncanplay assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.oncanplay assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.oncanplay assert_true: expected true got undefined
FAIL Reflect SVGElement.oncanplay assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.oncanplay assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.oncanplay assert_true: expected true got undefined
FAIL Set Window.oncanplay assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.oncanplay 
PASS No Element.oncanplaythrough 
FAIL Set HTMLElement.oncanplaythrough assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.oncanplaythrough assert_true: expected true got undefined
FAIL Reflect HTMLElement.oncanplaythrough assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.oncanplaythrough assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.oncanplaythrough assert_true: expected true got undefined
FAIL Reflect SVGElement.oncanplaythrough assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.oncanplaythrough assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.oncanplaythrough assert_true: expected true got undefined
FAIL Set Window.oncanplaythrough assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.oncanplaythrough 
FAIL No Element.onchange assert_false: expected false got true
PASS Set HTMLElement.onchange 
PASS Enumerate HTMLElement.onchange 
PASS Reflect HTMLElement.onchange 
PASS Set SVGElement.onchange 
PASS Enumerate SVGElement.onchange 
FAIL Reflect SVGElement.onchange assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onchange 
PASS Enumerate Document.onchange 
PASS Set Window.onchange 
PASS Enumerate Window.onchange 
FAIL No Element.onclick assert_false: expected false got true
PASS Set HTMLElement.onclick 
PASS Enumerate HTMLElement.onclick 
PASS Reflect HTMLElement.onclick 
PASS Set SVGElement.onclick 
PASS Enumerate SVGElement.onclick 
PASS Reflect SVGElement.onclick 
PASS Set Document.onclick 
PASS Enumerate Document.onclick 
PASS Set Window.onclick 
PASS Enumerate Window.onclick 
PASS No Element.onclose 
FAIL Set HTMLElement.onclose assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onclose assert_true: expected true got undefined
FAIL Reflect HTMLElement.onclose assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onclose assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onclose assert_true: expected true got undefined
FAIL Reflect SVGElement.onclose assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onclose assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onclose assert_true: expected true got undefined
FAIL Set Window.onclose assert_equals: Initially null expected null but got undefined
FAIL Enumerate Window.onclose assert_true: expected true got undefined
FAIL No Element.oncontextmenu assert_false: expected false got true
PASS Set HTMLElement.oncontextmenu 
PASS Enumerate HTMLElement.oncontextmenu 
PASS Reflect HTMLElement.oncontextmenu 
PASS Set SVGElement.oncontextmenu 
PASS Enumerate SVGElement.oncontextmenu 
FAIL Reflect SVGElement.oncontextmenu assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.oncontextmenu 
PASS Enumerate Document.oncontextmenu 
PASS Set Window.oncontextmenu 
PASS Enumerate Window.oncontextmenu 
PASS No Element.oncuechange 
FAIL Set HTMLElement.oncuechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.oncuechange assert_true: expected true got undefined
FAIL Reflect HTMLElement.oncuechange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.oncuechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.oncuechange assert_true: expected true got undefined
FAIL Reflect SVGElement.oncuechange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.oncuechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.oncuechange assert_true: expected true got undefined
FAIL Set Window.oncuechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate Window.oncuechange assert_true: expected true got undefined
FAIL No Element.ondblclick assert_false: expected false got true
PASS Set HTMLElement.ondblclick 
PASS Enumerate HTMLElement.ondblclick 
PASS Reflect HTMLElement.ondblclick 
PASS Set SVGElement.ondblclick 
PASS Enumerate SVGElement.ondblclick 
FAIL Reflect SVGElement.ondblclick assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondblclick 
PASS Enumerate Document.ondblclick 
PASS Set Window.ondblclick 
PASS Enumerate Window.ondblclick 
FAIL No Element.ondrag assert_false: expected false got true
PASS Set HTMLElement.ondrag 
PASS Enumerate HTMLElement.ondrag 
PASS Reflect HTMLElement.ondrag 
PASS Set SVGElement.ondrag 
PASS Enumerate SVGElement.ondrag 
FAIL Reflect SVGElement.ondrag assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondrag 
PASS Enumerate Document.ondrag 
PASS Set Window.ondrag 
PASS Enumerate Window.ondrag 
FAIL No Element.ondragend assert_false: expected false got true
PASS Set HTMLElement.ondragend 
PASS Enumerate HTMLElement.ondragend 
PASS Reflect HTMLElement.ondragend 
PASS Set SVGElement.ondragend 
PASS Enumerate SVGElement.ondragend 
FAIL Reflect SVGElement.ondragend assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondragend 
PASS Enumerate Document.ondragend 
PASS Set Window.ondragend 
PASS Enumerate Window.ondragend 
FAIL No Element.ondragenter assert_false: expected false got true
PASS Set HTMLElement.ondragenter 
PASS Enumerate HTMLElement.ondragenter 
PASS Reflect HTMLElement.ondragenter 
PASS Set SVGElement.ondragenter 
PASS Enumerate SVGElement.ondragenter 
FAIL Reflect SVGElement.ondragenter assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondragenter 
PASS Enumerate Document.ondragenter 
PASS Set Window.ondragenter 
PASS Enumerate Window.ondragenter 
PASS No Element.ondragexit 
FAIL Set HTMLElement.ondragexit assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.ondragexit assert_true: expected true got undefined
FAIL Reflect HTMLElement.ondragexit assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.ondragexit assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.ondragexit assert_true: expected true got undefined
FAIL Reflect SVGElement.ondragexit assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.ondragexit assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.ondragexit assert_true: expected true got undefined
FAIL Set Window.ondragexit assert_equals: Initially null expected null but got undefined
FAIL Enumerate Window.ondragexit assert_true: expected true got undefined
FAIL No Element.ondragleave assert_false: expected false got true
PASS Set HTMLElement.ondragleave 
PASS Enumerate HTMLElement.ondragleave 
PASS Reflect HTMLElement.ondragleave 
PASS Set SVGElement.ondragleave 
PASS Enumerate SVGElement.ondragleave 
FAIL Reflect SVGElement.ondragleave assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondragleave 
PASS Enumerate Document.ondragleave 
PASS Set Window.ondragleave 
PASS Enumerate Window.ondragleave 
FAIL No Element.ondragover assert_false: expected false got true
PASS Set HTMLElement.ondragover 
PASS Enumerate HTMLElement.ondragover 
PASS Reflect HTMLElement.ondragover 
PASS Set SVGElement.ondragover 
PASS Enumerate SVGElement.ondragover 
FAIL Reflect SVGElement.ondragover assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondragover 
PASS Enumerate Document.ondragover 
PASS Set Window.ondragover 
PASS Enumerate Window.ondragover 
FAIL No Element.ondragstart assert_false: expected false got true
PASS Set HTMLElement.ondragstart 
PASS Enumerate HTMLElement.ondragstart 
PASS Reflect HTMLElement.ondragstart 
PASS Set SVGElement.ondragstart 
PASS Enumerate SVGElement.ondragstart 
FAIL Reflect SVGElement.ondragstart assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondragstart 
PASS Enumerate Document.ondragstart 
PASS Set Window.ondragstart 
PASS Enumerate Window.ondragstart 
FAIL No Element.ondrop assert_false: expected false got true
PASS Set HTMLElement.ondrop 
PASS Enumerate HTMLElement.ondrop 
PASS Reflect HTMLElement.ondrop 
PASS Set SVGElement.ondrop 
PASS Enumerate SVGElement.ondrop 
FAIL Reflect SVGElement.ondrop assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.ondrop 
PASS Enumerate Document.ondrop 
PASS Set Window.ondrop 
PASS Enumerate Window.ondrop 
PASS No Element.ondurationchange 
FAIL Set HTMLElement.ondurationchange assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.ondurationchange assert_true: expected true got undefined
FAIL Reflect HTMLElement.ondurationchange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.ondurationchange assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.ondurationchange assert_true: expected true got undefined
FAIL Reflect SVGElement.ondurationchange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.ondurationchange assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.ondurationchange assert_true: expected true got undefined
FAIL Set Window.ondurationchange assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.ondurationchange 
PASS No Element.onemptied 
FAIL Set HTMLElement.onemptied assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onemptied assert_true: expected true got undefined
FAIL Reflect HTMLElement.onemptied assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onemptied assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onemptied assert_true: expected true got undefined
FAIL Reflect SVGElement.onemptied assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onemptied assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onemptied assert_true: expected true got undefined
FAIL Set Window.onemptied assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onemptied 
PASS No Element.onended 
FAIL Set HTMLElement.onended assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onended assert_true: expected true got undefined
FAIL Reflect HTMLElement.onended assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onended assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onended assert_true: expected true got undefined
FAIL Reflect SVGElement.onended assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onended assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onended assert_true: expected true got undefined
FAIL Set Window.onended assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onended 
FAIL No Element.oninput assert_false: expected false got true
PASS Set HTMLElement.oninput 
PASS Enumerate HTMLElement.oninput 
PASS Reflect HTMLElement.oninput 
PASS Set SVGElement.oninput 
PASS Enumerate SVGElement.oninput 
FAIL Reflect SVGElement.oninput assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.oninput 
PASS Enumerate Document.oninput 
PASS Set Window.oninput 
PASS Enumerate Window.oninput 
FAIL No Element.oninvalid assert_false: expected false got true
PASS Set HTMLElement.oninvalid 
PASS Enumerate HTMLElement.oninvalid 
PASS Reflect HTMLElement.oninvalid 
PASS Set SVGElement.oninvalid 
PASS Enumerate SVGElement.oninvalid 
FAIL Reflect SVGElement.oninvalid assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.oninvalid 
PASS Enumerate Document.oninvalid 
PASS Set Window.oninvalid 
PASS Enumerate Window.oninvalid 
FAIL No Element.onkeydown assert_false: expected false got true
PASS Set HTMLElement.onkeydown 
PASS Enumerate HTMLElement.onkeydown 
PASS Reflect HTMLElement.onkeydown 
PASS Set SVGElement.onkeydown 
PASS Enumerate SVGElement.onkeydown 
FAIL Reflect SVGElement.onkeydown assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onkeydown 
PASS Enumerate Document.onkeydown 
PASS Set Window.onkeydown 
PASS Enumerate Window.onkeydown 
FAIL No Element.onkeypress assert_false: expected false got true
PASS Set HTMLElement.onkeypress 
PASS Enumerate HTMLElement.onkeypress 
PASS Reflect HTMLElement.onkeypress 
PASS Set SVGElement.onkeypress 
PASS Enumerate SVGElement.onkeypress 
FAIL Reflect SVGElement.onkeypress assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onkeypress 
PASS Enumerate Document.onkeypress 
PASS Set Window.onkeypress 
PASS Enumerate Window.onkeypress 
FAIL No Element.onkeyup assert_false: expected false got true
PASS Set HTMLElement.onkeyup 
PASS Enumerate HTMLElement.onkeyup 
PASS Reflect HTMLElement.onkeyup 
PASS Set SVGElement.onkeyup 
PASS Enumerate SVGElement.onkeyup 
FAIL Reflect SVGElement.onkeyup assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onkeyup 
PASS Enumerate Document.onkeyup 
PASS Set Window.onkeyup 
PASS Enumerate Window.onkeyup 
FAIL No Element.onload assert_false: expected false got true
PASS Set HTMLElement.onload 
PASS Enumerate HTMLElement.onload 
PASS Reflect HTMLElement.onload 
PASS Set SVGElement.onload 
PASS Enumerate SVGElement.onload 
PASS Reflect SVGElement.onload 
PASS Set Document.onload 
PASS Enumerate Document.onload 
PASS Set Window.onload 
PASS Enumerate Window.onload 
PASS No Element.onloadeddata 
FAIL Set HTMLElement.onloadeddata assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onloadeddata assert_true: expected true got undefined
FAIL Reflect HTMLElement.onloadeddata assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onloadeddata assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onloadeddata assert_true: expected true got undefined
FAIL Reflect SVGElement.onloadeddata assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onloadeddata assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onloadeddata assert_true: expected true got undefined
FAIL Set Window.onloadeddata assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onloadeddata 
PASS No Element.onloadedmetadata 
FAIL Set HTMLElement.onloadedmetadata assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onloadedmetadata assert_true: expected true got undefined
FAIL Reflect HTMLElement.onloadedmetadata assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onloadedmetadata assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onloadedmetadata assert_true: expected true got undefined
FAIL Reflect SVGElement.onloadedmetadata assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onloadedmetadata assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onloadedmetadata assert_true: expected true got undefined
FAIL Set Window.onloadedmetadata assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onloadedmetadata 
PASS No Element.onloadstart 
FAIL Set HTMLElement.onloadstart assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onloadstart assert_true: expected true got undefined
FAIL Reflect HTMLElement.onloadstart assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onloadstart assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onloadstart assert_true: expected true got undefined
FAIL Reflect SVGElement.onloadstart assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onloadstart assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onloadstart assert_true: expected true got undefined
FAIL Set Window.onloadstart assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onloadstart 
FAIL No Element.onmousedown assert_false: expected false got true
PASS Set HTMLElement.onmousedown 
PASS Enumerate HTMLElement.onmousedown 
PASS Reflect HTMLElement.onmousedown 
PASS Set SVGElement.onmousedown 
PASS Enumerate SVGElement.onmousedown 
PASS Reflect SVGElement.onmousedown 
PASS Set Document.onmousedown 
PASS Enumerate Document.onmousedown 
PASS Set Window.onmousedown 
PASS Enumerate Window.onmousedown 
FAIL No Element.onmouseenter assert_false: expected false got true
PASS Set HTMLElement.onmouseenter 
PASS Enumerate HTMLElement.onmouseenter 
PASS Reflect HTMLElement.onmouseenter 
PASS Set SVGElement.onmouseenter 
PASS Enumerate SVGElement.onmouseenter 
PASS Reflect SVGElement.onmouseenter 
PASS Set Document.onmouseenter 
PASS Enumerate Document.onmouseenter 
PASS Set Window.onmouseenter 
PASS Enumerate Window.onmouseenter 
FAIL No Element.onmouseleave assert_false: expected false got true
PASS Set HTMLElement.onmouseleave 
PASS Enumerate HTMLElement.onmouseleave 
PASS Reflect HTMLElement.onmouseleave 
PASS Set SVGElement.onmouseleave 
PASS Enumerate SVGElement.onmouseleave 
PASS Reflect SVGElement.onmouseleave 
PASS Set Document.onmouseleave 
PASS Enumerate Document.onmouseleave 
PASS Set Window.onmouseleave 
PASS Enumerate Window.onmouseleave 
FAIL No Element.onmousemove assert_false: expected false got true
PASS Set HTMLElement.onmousemove 
PASS Enumerate HTMLElement.onmousemove 
PASS Reflect HTMLElement.onmousemove 
PASS Set SVGElement.onmousemove 
PASS Enumerate SVGElement.onmousemove 
PASS Reflect SVGElement.onmousemove 
PASS Set Document.onmousemove 
PASS Enumerate Document.onmousemove 
PASS Set Window.onmousemove 
PASS Enumerate Window.onmousemove 
FAIL No Element.onmouseout assert_false: expected false got true
PASS Set HTMLElement.onmouseout 
PASS Enumerate HTMLElement.onmouseout 
PASS Reflect HTMLElement.onmouseout 
PASS Set SVGElement.onmouseout 
PASS Enumerate SVGElement.onmouseout 
PASS Reflect SVGElement.onmouseout 
PASS Set Document.onmouseout 
PASS Enumerate Document.onmouseout 
PASS Set Window.onmouseout 
PASS Enumerate Window.onmouseout 
FAIL No Element.onmouseover assert_false: expected false got true
PASS Set HTMLElement.onmouseover 
PASS Enumerate HTMLElement.onmouseover 
PASS Reflect HTMLElement.onmouseover 
PASS Set SVGElement.onmouseover 
PASS Enumerate SVGElement.onmouseover 
PASS Reflect SVGElement.onmouseover 
PASS Set Document.onmouseover 
PASS Enumerate Document.onmouseover 
PASS Set Window.onmouseover 
PASS Enumerate Window.onmouseover 
FAIL No Element.onmouseup assert_false: expected false got true
PASS Set HTMLElement.onmouseup 
PASS Enumerate HTMLElement.onmouseup 
PASS Reflect HTMLElement.onmouseup 
PASS Set SVGElement.onmouseup 
PASS Enumerate SVGElement.onmouseup 
PASS Reflect SVGElement.onmouseup 
PASS Set Document.onmouseup 
PASS Enumerate Document.onmouseup 
PASS Set Window.onmouseup 
PASS Enumerate Window.onmouseup 
FAIL No Element.onmousewheel assert_false: expected false got true
PASS Set HTMLElement.onmousewheel 
PASS Enumerate HTMLElement.onmousewheel 
PASS Reflect HTMLElement.onmousewheel 
PASS Set SVGElement.onmousewheel 
PASS Enumerate SVGElement.onmousewheel 
FAIL Reflect SVGElement.onmousewheel assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onmousewheel 
PASS Enumerate Document.onmousewheel 
PASS Set Window.onmousewheel 
PASS Enumerate Window.onmousewheel 
PASS No Element.onpause 
FAIL Set HTMLElement.onpause assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onpause assert_true: expected true got undefined
FAIL Reflect HTMLElement.onpause assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onpause assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onpause assert_true: expected true got undefined
FAIL Reflect SVGElement.onpause assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onpause assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onpause assert_true: expected true got undefined
FAIL Set Window.onpause assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onpause 
PASS No Element.onplay 
FAIL Set HTMLElement.onplay assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onplay assert_true: expected true got undefined
FAIL Reflect HTMLElement.onplay assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onplay assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onplay assert_true: expected true got undefined
FAIL Reflect SVGElement.onplay assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onplay assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onplay assert_true: expected true got undefined
FAIL Set Window.onplay assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onplay 
PASS No Element.onplaying 
FAIL Set HTMLElement.onplaying assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onplaying assert_true: expected true got undefined
FAIL Reflect HTMLElement.onplaying assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onplaying assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onplaying assert_true: expected true got undefined
FAIL Reflect SVGElement.onplaying assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onplaying assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onplaying assert_true: expected true got undefined
FAIL Set Window.onplaying assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onplaying 
PASS No Element.onprogress 
FAIL Set HTMLElement.onprogress assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onprogress assert_true: expected true got undefined
FAIL Reflect HTMLElement.onprogress assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onprogress assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onprogress assert_true: expected true got undefined
FAIL Reflect SVGElement.onprogress assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onprogress assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onprogress assert_true: expected true got undefined
FAIL Set Window.onprogress assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onprogress 
PASS No Element.onratechange 
FAIL Set HTMLElement.onratechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onratechange assert_true: expected true got undefined
FAIL Reflect HTMLElement.onratechange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onratechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onratechange assert_true: expected true got undefined
FAIL Reflect SVGElement.onratechange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onratechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onratechange assert_true: expected true got undefined
FAIL Set Window.onratechange assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onratechange 
FAIL No Element.onreset assert_false: expected false got true
PASS Set HTMLElement.onreset 
PASS Enumerate HTMLElement.onreset 
PASS Reflect HTMLElement.onreset 
PASS Set SVGElement.onreset 
PASS Enumerate SVGElement.onreset 
FAIL Reflect SVGElement.onreset assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onreset 
PASS Enumerate Document.onreset 
PASS Set Window.onreset 
PASS Enumerate Window.onreset 
FAIL No Element.onscroll assert_false: expected false got true
PASS Set HTMLElement.onscroll 
PASS Enumerate HTMLElement.onscroll 
PASS Reflect HTMLElement.onscroll 
PASS Set SVGElement.onscroll 
PASS Enumerate SVGElement.onscroll 
FAIL Reflect SVGElement.onscroll assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onscroll 
PASS Enumerate Document.onscroll 
PASS Set Window.onscroll 
PASS Enumerate Window.onscroll 
PASS No Element.onseeked 
FAIL Set HTMLElement.onseeked assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onseeked assert_true: expected true got undefined
FAIL Reflect HTMLElement.onseeked assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onseeked assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onseeked assert_true: expected true got undefined
FAIL Reflect SVGElement.onseeked assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onseeked assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onseeked assert_true: expected true got undefined
FAIL Set Window.onseeked assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onseeked 
PASS No Element.onseeking 
FAIL Set HTMLElement.onseeking assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onseeking assert_true: expected true got undefined
FAIL Reflect HTMLElement.onseeking assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onseeking assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onseeking assert_true: expected true got undefined
FAIL Reflect SVGElement.onseeking assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onseeking assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onseeking assert_true: expected true got undefined
FAIL Set Window.onseeking assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onseeking 
FAIL No Element.onselect assert_false: expected false got true
PASS Set HTMLElement.onselect 
PASS Enumerate HTMLElement.onselect 
PASS Reflect HTMLElement.onselect 
PASS Set SVGElement.onselect 
PASS Enumerate SVGElement.onselect 
FAIL Reflect SVGElement.onselect assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onselect 
PASS Enumerate Document.onselect 
PASS Set Window.onselect 
PASS Enumerate Window.onselect 
PASS No Element.onshow 
FAIL Set HTMLElement.onshow assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onshow assert_true: expected true got undefined
FAIL Reflect HTMLElement.onshow assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onshow assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onshow assert_true: expected true got undefined
FAIL Reflect SVGElement.onshow assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onshow assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onshow assert_true: expected true got undefined
FAIL Set Window.onshow assert_equals: Initially null expected null but got undefined
FAIL Enumerate Window.onshow assert_true: expected true got undefined
PASS No Element.onsort 
FAIL Set HTMLElement.onsort assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onsort assert_true: expected true got undefined
FAIL Reflect HTMLElement.onsort assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onsort assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onsort assert_true: expected true got undefined
FAIL Reflect SVGElement.onsort assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onsort assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onsort assert_true: expected true got undefined
FAIL Set Window.onsort assert_equals: Initially null expected null but got undefined
FAIL Enumerate Window.onsort assert_true: expected true got undefined
PASS No Element.onstalled 
FAIL Set HTMLElement.onstalled assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onstalled assert_true: expected true got undefined
FAIL Reflect HTMLElement.onstalled assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onstalled assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onstalled assert_true: expected true got undefined
FAIL Reflect SVGElement.onstalled assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onstalled assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onstalled assert_true: expected true got undefined
FAIL Set Window.onstalled assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onstalled 
FAIL No Element.onsubmit assert_false: expected false got true
PASS Set HTMLElement.onsubmit 
PASS Enumerate HTMLElement.onsubmit 
PASS Reflect HTMLElement.onsubmit 
PASS Set SVGElement.onsubmit 
PASS Enumerate SVGElement.onsubmit 
FAIL Reflect SVGElement.onsubmit assert_equals: Convert to function expected "function" but got "object"
PASS Set Document.onsubmit 
PASS Enumerate Document.onsubmit 
PASS Set Window.onsubmit 
PASS Enumerate Window.onsubmit 
PASS No Element.onsuspend 
FAIL Set HTMLElement.onsuspend assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onsuspend assert_true: expected true got undefined
FAIL Reflect HTMLElement.onsuspend assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onsuspend assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onsuspend assert_true: expected true got undefined
FAIL Reflect SVGElement.onsuspend assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onsuspend assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onsuspend assert_true: expected true got undefined
FAIL Set Window.onsuspend assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onsuspend 
PASS No Element.ontimeupdate 
FAIL Set HTMLElement.ontimeupdate assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.ontimeupdate assert_true: expected true got undefined
FAIL Reflect HTMLElement.ontimeupdate assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.ontimeupdate assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.ontimeupdate assert_true: expected true got undefined
FAIL Reflect SVGElement.ontimeupdate assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.ontimeupdate assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.ontimeupdate assert_true: expected true got undefined
FAIL Set Window.ontimeupdate assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.ontimeupdate 
PASS No Element.onvolumechange 
FAIL Set HTMLElement.onvolumechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onvolumechange assert_true: expected true got undefined
FAIL Reflect HTMLElement.onvolumechange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onvolumechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onvolumechange assert_true: expected true got undefined
FAIL Reflect SVGElement.onvolumechange assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onvolumechange assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onvolumechange assert_true: expected true got undefined
FAIL Set Window.onvolumechange assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onvolumechange 
PASS No Element.onwaiting 
FAIL Set HTMLElement.onwaiting assert_equals: Initially null expected null but got undefined
FAIL Enumerate HTMLElement.onwaiting assert_true: expected true got undefined
FAIL Reflect HTMLElement.onwaiting assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set SVGElement.onwaiting assert_equals: Initially null expected null but got undefined
FAIL Enumerate SVGElement.onwaiting assert_true: expected true got undefined
FAIL Reflect SVGElement.onwaiting assert_equals: Convert to function expected "function" but got "undefined"
FAIL Set Document.onwaiting assert_equals: Initially null expected null but got undefined
FAIL Enumerate Document.onwaiting assert_true: expected true got undefined
FAIL Set Window.onwaiting assert_equals: Return null after setting string expected null but got ""
PASS Enumerate Window.onwaiting 
